as_esaps

Convierte salida de la funcion nacional_uy al formato necesario para poder usar el paquete esaps.

as_esaps(datos)

Arguments

datos
Datos de la clase boreluy.

Value

data.frame.

Examples

elec <- nacional_uy(eleccion = 1971)

elec <- as_esaps(elec)

# Numero efectivo de partidos
# esaps::enp(elec)

# Desproporcionalidad electoral
# esaps::dispro(elec, method = 1)

# Nacionalizacion del sistema de partidos y de los partidos politicos
# esaps::psns(tidy_data = elec, method = 1, pns = TRUE)