as_parliament

Convierte salida de la funcion nacional_uy al formato necesario para poder usar el paquete ggparliament.

as_parliament(datos, camara = 1, color, tipo = "semicircle", ...)

Arguments

datos
Datos de la clase boreluy.

camara
Por defecto es 1 que correponde a la camara baja o Camara de Represnetantes, si se indica 2 corresponde a la camara alta o Camara de Senadores.

color
Paleta de colores con los que se van a identificar a los partidos politicos en el grafico.

tipo
Tipo de parlamento (horseshoe, semicircle, circle, classroom, opposing benches)

...
argumentos adicionales de la funcion parliament_data

Value

data.frame.

Examples

diputados <- nacional_uy(eleccion = 1971)
as_parliament(diputados, camara = 1, color = c('#E81B23', '#3333FF', '#B4B4B4'))

#>      year country          house       party_long party_short seats government
#> 1    1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.1  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.2  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.3  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.4  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.5  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.6  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.7  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.8  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.9  1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.10 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.11 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.12 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.13 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.14 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.15 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.16 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.17 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.18 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.19 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.20 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.21 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.22 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.23 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.24 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.25 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.26 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.27 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.28 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.29 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.30 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.31 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.32 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.33 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.34 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.35 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.36 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.37 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.38 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.39 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 1.40 1971 Uruguay Representantes Partido Colorado          PC    41          1
#> 2    1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.1  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.2  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.3  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.4  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.5  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.6  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.7  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.8  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.9  1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.10 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.11 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.12 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.13 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.14 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.15 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.16 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.17 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.18 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.19 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.20 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.21 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.22 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.23 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.24 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.25 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.26 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.27 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.28 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.29 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.30 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.31 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.32 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.33 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.34 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.35 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.36 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.37 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.38 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 2.39 1971 Uruguay Representantes Partido Nacional          PN    40          0
#> 3    1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.1  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.2  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.3  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.4  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.5  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.6  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.7  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.8  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.9  1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.10 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.11 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.12 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.13 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.14 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.15 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.16 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#> 3.17 1971 Uruguay Representantes    Frente Amplio          FA    18          0
#>                colour             x            y row      theta
#> 1              gray20 -2.000000e+00 2.449213e-16   3 3.14159265
#> 1.1            gray20 -1.500000e+00 1.836910e-16   2 3.14159265
#> 1.2            gray20 -1.000000e+00 1.224606e-16   1 3.14159265
#> 1.3            gray20 -1.994665e+00 1.459906e-01   3 3.06853236
#> 1.4            gray20 -1.492777e+00 1.470257e-01   2 3.04341788
#> 1.5            gray20 -1.978687e+00 2.912023e-01   3 2.99547207
#> 1.6            gray20 -9.888308e-01 1.490423e-01   1 2.99199300
#> 1.7            gray20 -1.471178e+00 2.926355e-01   2 2.94524311
#> 1.8            gray20 -1.952152e+00 4.348604e-01   3 2.92241177
#> 1.9            gray20 -1.915201e+00 5.761982e-01   3 2.84935148
#> 1.10           gray20 -1.435411e+00 4.354270e-01   2 2.84706834
#> 1.11           gray20 -9.555728e-01 2.947552e-01   1 2.84239335
#> 1.12           gray20 -1.868032e+00 7.144618e-01   3 2.77629118
#> 1.13           gray20 -1.385819e+00 5.740251e-01   2 2.74889357
#> 1.14           gray20 -1.810896e+00 8.489134e-01   3 2.70323089
#> 1.15           gray20 -9.009689e-01 4.338837e-01   1 2.69279370
#> 1.16           gray20 -1.322882e+00 7.070951e-01   2 2.65071880
#> 1.17           gray20 -1.744099e+00 9.788357e-01   3 2.63017059
#> 1.18           gray20 -1.667996e+00 1.103535e+00   3 2.55711030
#> 1.19           gray20 -1.247204e+00 8.333553e-01   2 2.55254403
#> 1.20           gray20 -8.262388e-01 5.633201e-01   1 2.54319405
#> 1.21           gray20 -1.582993e+00 1.222347e+00   3 2.48405001
#> 1.22           gray20 -1.159516e+00 9.515899e-01   2 2.45436926
#> 1.23           gray20 -1.489544e+00 1.334638e+00   3 2.41098971
#> 1.24           gray20 -7.330519e-01 6.801727e-01   1 2.39359440
#> 1.25           gray20 -1.060660e+00 1.060660e+00   2 2.35619449
#> 1.26           gray20 -1.388148e+00 1.439807e+00   3 2.33792942
#> 1.27           gray20 -1.279346e+00 1.537294e+00   3 2.26486912
#> 1.28           gray20 -9.515899e-01 1.159516e+00   2 2.25801972
#> 1.29           gray20 -6.234898e-01 7.818315e-01   1 2.24399475
#> 1.30           gray20 -1.163718e+00 1.626579e+00   3 2.19180883
#> 1.31           gray20 -8.333553e-01 1.247204e+00   2 2.15984495
#> 1.32           gray20 -1.041881e+00 1.707186e+00   3 2.11874853
#> 1.33           gray20 -5.000000e-01 8.660254e-01   1 2.09439510
#> 1.34           gray20 -7.070951e-01 1.322882e+00   2 2.06167018
#> 1.35           gray20 -9.144846e-01 1.778684e+00   3 2.04568824
#> 1.36           gray20 -7.822094e-01 1.840692e+00   3 1.97262795
#> 1.37           gray20 -5.740251e-01 1.385819e+00   2 1.96349541
#> 1.38           gray20 -3.653410e-01 9.308737e-01   1 1.94479545
#> 1.39           gray20 -6.457608e-01 1.892880e+00   3 1.89956765
#> 1.40           gray20 -4.354270e-01 1.435411e+00   2 1.86532064
#> 2          lightcyan2 -5.058668e-01 1.934967e+00   3 1.82650736
#> 2.1        lightcyan2 -2.225209e-01 9.749279e-01   1 1.79519580
#> 2.2        lightcyan2 -2.926355e-01 1.471178e+00   2 1.76714587
#> 2.3        lightcyan2 -3.632737e-01 1.966731e+00   3 1.75344706
#> 2.4        lightcyan2 -2.187424e-01 1.988002e+00   3 1.68038677
#> 2.5        lightcyan2 -1.470257e-01 1.492777e+00   2 1.66897110
#> 2.6        lightcyan2 -7.473009e-02 9.972038e-01   1 1.64559615
#> 2.7        lightcyan2 -7.304405e-02 1.998666e+00   3 1.60732647
#> 2.8        lightcyan2  9.184548e-17 1.500000e+00   2 1.57079633
#> 2.9        lightcyan2  7.304405e-02 1.998666e+00   3 1.53426618
#> 2.10       lightcyan2  7.473009e-02 9.972038e-01   1 1.49599650
#> 2.11       lightcyan2  1.470257e-01 1.492777e+00   2 1.47262156
#> 2.12       lightcyan2  2.187424e-01 1.988002e+00   3 1.46120589
#> 2.13       lightcyan2  3.632737e-01 1.966731e+00   3 1.38814559
#> 2.14       lightcyan2  2.926355e-01 1.471178e+00   2 1.37444679
#> 2.15       lightcyan2  2.225209e-01 9.749279e-01   1 1.34639685
#> 2.16       lightcyan2  5.058668e-01 1.934967e+00   3 1.31508530
#> 2.17       lightcyan2  4.354270e-01 1.435411e+00   2 1.27627202
#> 2.18       lightcyan2  6.457608e-01 1.892880e+00   3 1.24202500
#> 2.19       lightcyan2  3.653410e-01 9.308737e-01   1 1.19679720
#> 2.20       lightcyan2  5.740251e-01 1.385819e+00   2 1.17809725
#> 2.21       lightcyan2  7.822094e-01 1.840692e+00   3 1.16896471
#> 2.22       lightcyan2  9.144846e-01 1.778684e+00   3 1.09590441
#> 2.23       lightcyan2  7.070951e-01 1.322882e+00   2 1.07992247
#> 2.24       lightcyan2  5.000000e-01 8.660254e-01   1 1.04719755
#> 2.25       lightcyan2  1.041881e+00 1.707186e+00   3 1.02284412
#> 2.26       lightcyan2  8.333553e-01 1.247204e+00   2 0.98174770
#> 2.27       lightcyan2  1.163718e+00 1.626579e+00   3 0.94978383
#> 2.28       lightcyan2  6.234898e-01 7.818315e-01   1 0.89759790
#> 2.29       lightcyan2  9.515899e-01 1.159516e+00   2 0.88357293
#> 2.30       lightcyan2  1.279346e+00 1.537294e+00   3 0.87672353
#> 2.31       lightcyan2  1.388148e+00 1.439807e+00   3 0.80366324
#> 2.32       lightcyan2  1.060660e+00 1.060660e+00   2 0.78539816
#> 2.33       lightcyan2  7.330519e-01 6.801727e-01   1 0.74799825
#> 2.34       lightcyan2  1.489544e+00 1.334638e+00   3 0.73060294
#> 2.35       lightcyan2  1.159516e+00 9.515899e-01   2 0.68722339
#> 2.36       lightcyan2  1.582993e+00 1.222347e+00   3 0.65754265
#> 2.37       lightcyan2  8.262388e-01 5.633201e-01   1 0.59839860
#> 2.38       lightcyan2  1.247204e+00 8.333553e-01   2 0.58904862
#> 2.39       lightcyan2  1.667996e+00 1.103535e+00   3 0.58448235
#> 3    mediumaquamarine  1.744099e+00 9.788357e-01   3 0.51142206
#> 3.1  mediumaquamarine  1.322882e+00 7.070951e-01   2 0.49087385
#> 3.2  mediumaquamarine  9.009689e-01 4.338837e-01   1 0.44879895
#> 3.3  mediumaquamarine  1.810896e+00 8.489134e-01   3 0.43836177
#> 3.4  mediumaquamarine  1.385819e+00 5.740251e-01   2 0.39269908
#> 3.5  mediumaquamarine  1.868032e+00 7.144618e-01   3 0.36530147
#> 3.6  mediumaquamarine  9.555728e-01 2.947552e-01   1 0.29919930
#> 3.7  mediumaquamarine  1.435411e+00 4.354270e-01   2 0.29452431
#> 3.8  mediumaquamarine  1.915201e+00 5.761982e-01   3 0.29224118
#> 3.9  mediumaquamarine  1.952152e+00 4.348604e-01   3 0.21918088
#> 3.10 mediumaquamarine  1.471178e+00 2.926355e-01   2 0.19634954
#> 3.11 mediumaquamarine  9.888308e-01 1.490423e-01   1 0.14959965
#> 3.12 mediumaquamarine  1.978687e+00 2.912023e-01   3 0.14612059
#> 3.13 mediumaquamarine  1.492777e+00 1.470257e-01   2 0.09817477
#> 3.14 mediumaquamarine  1.994665e+00 1.459906e-01   3 0.07306029
#> 3.15 mediumaquamarine  2.000000e+00 0.000000e+00   3 0.00000000
#> 3.16 mediumaquamarine  1.500000e+00 0.000000e+00   2 0.00000000
#> 3.17 mediumaquamarine  1.000000e+00 0.000000e+00   1 0.00000000

# ggplot2::ggplot(diputados, aes(x, y, colour = party_long)) +
#    ggplot2::geom_parliament_seats() +
#    ggplot2::geom_highlight_government(government == 1) +
#    ggplot2::draw_majoritythreshold(n = mayoria, label = FALSE, type = 'semicircle') +
#    ggplot2::geom_parliament_bar(colour = colour, party = party_long) +
#    ggplot2::theme_ggparliament() +
#    ggplot2::labs(colour = NULL, title = titulo, subtitle = "") +
#    ggplot2::scale_colour_manual(values = datos$colour, limits = datos$party_long)