elecciones_uy

Contiene el resultado de un conjunto de elecciones de nivel nacional de Uruaguat entre 1918 y 2020.

data(elecciones_uy)

Format

data.frame con 18332 filas y 24 columnas.

Details

anio_eleccion

Anio en el que se realizo la eleccion.

fecha

Fecha en la que se realizo la eleccion.

eleccion

Tipo de eleccion. Puede ser 'Presidencial', 'Departamental', 'Balotaje', 'Legislativa' o 'Consejo Nacional de Administracion'

concurrente

Elecciones concurrentes. Si hay elecciones concurrentes la variable asume el valor 1, en caso contrario asume el vaor de 0. La concurrencia en este caso solo refiere a elecciones Presidenciales y Legislativas.

parcial

Elecciones parciales. El parlamento en la decada de los anios 20 se cambiaba por tercios. Si la variable asume el valor de 1 es que es una eleccion parcial en caso contrario asume el valor de 0.

organo

Nombre del organo electivo.

cargos

Cantidad principal de cargos que se eligen en cada eleccion.

partido

Nombre del partido politico.

departamento

Nombre del departamento.

fraccion

Nombre de la fraccion del partido politico.

sublema_senado

Nombre del sublema al senado.

cabeza_senado

Nombre de la persona que encabeza la lista al senado.

votos

Cantidad de votos.

bancas_diputados

Cantidad de bancas a la Camara de Representantes obtenidas por cada partido politico.

bancas_senado

Cantidad de bancas a la Camara de Senadores obtenidas por cada partido politico.

total_bancas_diputados

Total de bancas que tiene la Camara de Representantes.

total_bancas_senado

Total de bancas que tiene la Camara de Senadores.

habilitados

Cantidad de ciudadanos habilitados para votar.

poblacion

Cantidad de poblacion (no está discriminada por lo mayores de edad que son los que pueden ejercer el derecho al voto dentro de otras condiciones).

formula_presidente

Candidato a la presidncia si es eleccion Presidencial o a Intendente si es eleccion Departamental

formula_vice

Candidato a la vicepresidencia en el caso de elecciones Presidenciales.

partido_ganador

Partido politico ganador de la eleccion.

presidente

Presidente electo en el caso de elecciones Presidenciales o Intendente en el caso de elecciones Departamentales.

vice

Vicepresidente electo en el caso de elecciones Presidenciales.

La unidad de analisis de esta base de datos es la fraccion a nivel departamental. Esto hace que se tenga el resultado electoral estructurado por cada fraccion de cada partido en cada departamento.

References

Fabregat JT (1950). Elecciones uruguayas: Febrero de 1925 a noviembre de 1946, volume 1. Republica Oriental del Uruguay, Poder Legislativo, Camara de Representantes.

Fabregat JT (1957). Elecciones uruguayas: Noviembre de 1950 a noviembre de 1954.. Camara de Senadores.

Fabregat JT (1962). Elecciones uruguayas: elecciones del 30 de noviembre de 1958; compilacion de cifras oficiales, escrutinios y proclamaciones. Republica Oriental del Uruguay, Camara de Senadores.

Fabregat JT (1964). Elecciones uruguayas: del 25 de noviembre de 1962.. Camara de Senadores.

Fabregat JT (1968). Elecciones uruguayas: Plebiscitos y elecciones de noviembre 27 de 1966, volume 7. Republica Oriental del Uruguay, Poder Legislativo, Camara de Representantes.

Fabregat JT (1972). Elecciones uruguayas: Plebiscitos y elecciones de noviembre 28 de 1971, volume 8. Republica Oriental del Uruguay, Poder Legislativo, Camara de Representantes.

Albornoz A (1989). Elecciones generales del 25 de noviembre de 1984: eleccion de gobiernos nacionales y departamentales. Resultados. Proclamacion de titulares y suplentes. Republica Oriental del Uruguay, Poder Legislativo, Camara de Representantes.

Albornoz A (1992). Elecciones 1989. Republica Oriental del Uruguay, Poder Legislativo, Camara de Representantes.

Marius JL (2004). Elecciones uruguayas: 1980-2003. Konrad Adenauer Stiftung.

Marius JL, Gimenez W (2007). Primera eleccion uruguaya del siglo XXI: 31 de octubre de 2004. Ediciones la Proa.

Bottinelli OO, Marius JL, Gimenez W (2010). Enciclopedia Electoral del Uruguay 1900-2010. Factum.