Resultado de elecciones nacionales de Uruguay

Devuelve el resultado electoral por eleccion nacional en cantdiad de votos y en porcentaje. Las eleccciones nacionales son: Presidencial, Balotaje, Legislativa y del Consejo Nacional de Administracion.

nacional_uy(
  eleccion,
  tipo = "Presidencial",
  por_departamento = FALSE,
  vbva.rm = FALSE
)

Arguments

eleccion

Anio de eleccion.

tipo

Tipo de eleccion. Por defecto es 'Presidencial'

por_departamento

Por defecto es FALSE. Si es TRUE devuelve el resultado por departamento.

vbva.rm

Eliminar del calculo a los votos en blanco y anulados. Esto puede ser util para elecciones como el 'Balotaje'.

Value

data.frame. Si la eleccion es concurrente (hay eleccion de cargos legislativos) devuelve la cantidad de bancas por partido. Si la eleccion no es concurrente solo devuelve el resultado en porcentaje de la eleccion por partido.

Examples

nacional_uy(eleccion = 1971)
#> # A tibble: 9 x 8
#>   Fecha      Eleccion     Partido    Sigla  Votos Porcentaje Diputados Senadores
#>   <date>     <chr>        <chr>      <chr>  <dbl>      <dbl>     <dbl>     <dbl>
#> 1 1971-11-28 Presidencial Partido C~ PC    681624   39.8            41        14
#> 2 1971-11-28 Presidencial Partido N~ PN    668822   39.0            40        12
#> 3 1971-11-28 Presidencial Frente Am~ FA    304275   17.8            18         5
#> 4 1971-11-28 Presidencial Voto Anul~ VA     42508    2.48            0         0
#> 5 1971-11-28 Presidencial Union Rad~ URC     8844    0.516           0         0
#> 6 1971-11-28 Presidencial Voto en B~ VB      6139    0.358           0         0
#> 7 1971-11-28 Presidencial Partido d~ PJP      288    0.0168          0         0
#> 8 1971-11-28 Presidencial Movimient~ MJ       241    0.0141          0         0
#> 9 1971-11-28 Presidencial Partido J~ PJDO      25    0.00146         0         0
nacional_uy(eleccion = 1999, por_departamento = TRUE)
#> # A tibble: 133 x 9
#>    Fecha      Eleccion    Departamento Partido Sigla  Votos Porcentaje Diputados
#>    <date>     <chr>       <chr>        <chr>   <chr>  <dbl>      <dbl>     <dbl>
#>  1 1999-10-31 Presidenci~ Artigas      Partid~ PC     19773      40.2          1
#>  2 1999-10-31 Presidenci~ Artigas      Partid~ PN     16595      33.8          1
#>  3 1999-10-31 Presidenci~ Artigas      Frente~ FA     10907      22.2          0
#>  4 1999-10-31 Presidenci~ Artigas      Voto e~ VB       686       1.4          0
#>  5 1999-10-31 Presidenci~ Artigas      Nuevo ~ NE       658       1.34         0
#>  6 1999-10-31 Presidenci~ Artigas      Voto A~ VA       472       0.96         0
#>  7 1999-10-31 Presidenci~ Artigas      Union ~ UC        35       0.07         0
#>  8 1999-10-31 Presidenci~ Canelones    Frente~ FA    106571      38.0          6
#>  9 1999-10-31 Presidenci~ Canelones    Partid~ PC     90355      32.2          4
#> 10 1999-10-31 Presidenci~ Canelones    Partid~ PN     59832      21.4          3
#> # ... with 123 more rows, and 1 more variable: Senadores <dbl>
nacional_uy(eleccion = 1999, tipo = 'Balotaje')
#> # A tibble: 2 x 6
#>   Fecha      Eleccion Partido          Sigla   Votos Porcentaje
#>   <date>     <chr>    <chr>            <chr>   <dbl>      <dbl>
#> 1 1999-11-28 Balotaje Partido Colorado PC    1158708       54.1
#> 2 1999-11-28 Balotaje Frente Amplio    FA     982049       45.9
#nacional_uy(1925, tipo = 'Legislativa')